server/requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
sqlalchemy
psycopg2-binary